tests: use bash option pipefail to fail if left-hand command failure
The test named "cqfd run with satisfied command requirements, using su"
checks if the cqfd launcher uses the su command in the container.

The docker image is built with the CQFD_DEBUG environment variable set
in order to enable the cqfd debug traces at runtime. Then, the launcher
outputs the backend used to revoke the privilege to the user builder
(uses either su or sudo).

The test pipes the cqfd output to search for the backend in use and it
exists with the appropriate return status. And it succeeds as the output
backend matches the expected one. But the cqfd launcher exits in error
after it outputs the trace.

The newly introduced su option --session-command (09a84ca cqfd: use su
option --session-command) is not supported by the su applet implemented
by busybox. Furthermore, that 2012's change from is supported by Ubuntu
since 20.04! See the test trace attached below.

It might be expected the test to fail if cqfd returns in error.

The default return status by a shell pipelines is given by the last
right-hand command of the pipeline: i.e. the successfull match for
launcher backend in this case.

The tests uses bash and that shell interpreter provides:
 - either the array variable $PIPESTATUS to set every command return
   statuses of a pipeline ($? but for every commands)
 - or the bash specific option pipefail to change the return status
   behavior of a pipeline. See below:

According to bash(1):

	Pipelines

	(...)

	The return status of a pipeline is the exit status of the last
	command, unless the pipefail option is enabled. If pipefail is
	enabled, the pipeline's return status is the value of the last
	(rightmost) command to exit with a non-zero status, or zero if
	all commands exit successfully. If the reserved word ! precedes
	a pipeline, the exit status of that pipeline is the logical
	negation of the exit status as described above. The shell waits
	for all commands in the pipeline to terminate before returning a
	value.

This sets the bash option pipefail to fail if any command of a pipeline
fails, which causes the test to fail if cqfd is not fixed.

Fixes:

	[14:07:32|notice] preparing "cqfd run with satisfied command requirements, using su"
	[+] Building 3.8s (7/7) FINISHED                                        docker:default
	 => [internal] load .dockerignore                                                 0.7s
	 => => transferring context: 2B                                                   0.0s
	 => [internal] load build definition from Dockerfile                              1.0s
	 => => transferring dockerfile: 125B                                              0.0s
	 => [internal] load metadata for docker.io/library/alpine:3.8                     1.2s
	 => [1/3] FROM docker.io/library/alpine:3.8@sha256:2bb501e6173d9d006e56de5bce272  0.0s
	 => CACHED [2/3] RUN apk update && apk upgrade                                    0.0s
	 => CACHED [3/3] RUN apk add bash shadow                                          0.0s
	 => exporting to image                                                            0.1s
	 => => exporting layers                                                           0.0s
	 => => writing image sha256:45dbd2be7ff2a7b863404f75380106acef99b6402de3360237aa  0.1s
	 => => naming to docker.io/library/cqfd_gportay_cqfd_test_8ce5476                 0.0s
	Creating mailbox file: No such file or directory
	debug: Using "su" to execute session command "true" as user "gportay"
	su: unrecognized option: session-command
	Usage: su [options] [LOGIN]

	Options:
	  -c, --command COMMAND         pass COMMAND to the invoked shell
	  -h, --help                    display this help message and exit
	  -, -l, --login                make the shell a login shell
	  -m, -p,
	  --preserve-environment        do not reset environment variables, and
	                                keep the same shell
	  -s, --shell SHELL             use SHELL instead of the default in passwd

	[14:07:40|info] result: cqfd run with satisfied command requirements, using su: PASS