cqfd: do not handle semicolon comments
The .ini files uses the semicolon for comments. Also, that exact same
character is one of the shell control operator for list of pipelines.

The hacky parser is not smart enough to make the distinction if a
semicolon is an .ini comment or if it is a shell control operator.

It considers every semicolon character as a .ini comment, and it removes
the remaining characters till the EOL (including the semicolon), even if
they are part of a list of pipeline.

As a consequence, it cuts off the list of pipelines if the command=
values contain a semicolon character, and it results in running the
left-hand only.

This stops to handle the semicolon comments to avoid amputating the list
of pipelines controled by semicolons that are set in values of the
command= properties.

Note: The parser converts the .ini file to bash; as a consequence, the
shell comments do not affect the parsing, as long as the hashmark does
not break the syntax **AND** if it not put at the end of the section
line.