cqfd: launcher: enable debug logs on --verbose
The script cqfd_launcher is created on host at every execution of cqfd;
it is run as the docker-run command.

The function debug() (that is duplicated in the generated launcher
script) escapes the dollar symbol of the variable CQFD_DEBUG.

Therefore, the test is evaluated at runtime in the container, and the
variable CQFD_DEBUG has to be set in the container to enable the debug
traces.

The variable CQFD_DEBUG is set by the host and it is never set in the
container environment.

As a consequence, the debug traces are not echo'ed even if cqfd is run
with the --verbose option.

This enables output of debug traces by removing the backslash that
escape the leading dollar symbol expanding the variable. This way, the
variable CQFD_DEBUG is expanded at buildtime by the command cat (instead
of exanded a runtime). Thus, it is hardcoded in the launcher script:

	debug() {
		test -n true && echo "debug: $*"
	}

Note: That solution is prefered over setting the variable CQFD_DEBUG in
the environment via --env CQFD_DEBUG that unecessarly polute the
container environment.