cqfd: add CQFD_DIR environment and update -d option
This adds the environment CQFD_DIR to set the path to the .cqfd/
directory. The path is either relative a absolute.

Important: The relative path is resolved using the working directory at
the moment the option -d is given; please mind about the option -C:

	cqfd                             # uses "$PWD/.cqfd"
	cqfd         -d cqfd             # uses "$PWD/cqfd"
	cqfd -C /tmp -d cqfd             # uses /tmp/cqfd
	cqfd         -d "$HOME/src/cqfd" # uses "$HOME/src/cqfd"
	cqfd -C /tmp -d "$HOME/src/cqfd" # uses "$HOME/src/cqfd"

The option -d is now using a PATH; it was previously using a (relative)
PATH component as it was concatenated to the project directory.

This way, this option seems more "standard" if it is compared to
--git-dir.

Fixes: #173